Update core.ts
When crawling a webpage, if a custom CSS selector is selected, but the initial page does not have a suitable CSS selection object, it will cause the crawler to fail. If a custom CSS selector is not used, the truly valuable page (which can filter out a lot of useless information through the CSS selector) will have a lot more useless information, such as the "homepage". So this submission can be used when a custom CSS selector is available, and when it is not possible, use "body" as the CSS selector